+++ tests/dummy/mirage/models/user.js
@@ -1,4 +1,6 @@
-import { Model } from 'ember-cli-mirage';
+import { Model, hasMany, belongsTo } from 'ember-cli-mirage';
 
 export default Model.extend({
+  friends: hasMany('user'),
+  bestie: belongsTo('user')
 });