+++ tests/dummy/mirage/serializers/user.js
@@ -1,0 +1,12 @@
+import BaseSerializer from './application';
+
+export default BaseSerializer.extend({
+  include: ['friends'],
+  links(model) {
+    return {
+      bestie: {
+        related: `/users/${model.bestieId}`
+      }
+    };
+  }
+});